% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hkio-data.R
\docType{data}
\name{hkio_users_df}
\alias{hkio_users_df}
\alias{hkio_tweets_df1}
\alias{hkio_tweets_df2}
\alias{hkio_tweets_df3}
\alias{hkio_tweets_df4}
\alias{hkio_tweets_df5}
\alias{hkio_tweets_df6}
\alias{hkio_tweets_df7}
\alias{hkio_tweets_df8}
\alias{hkio_tweets_df9}
\title{hkio}
\format{Ten \code{data.table}s. One for users, and nine covering the more than 3.6 million tweets.}
\source{
\url{https://blog.twitter.com/en_us/topics/company/2019/information_operations_directed_at_Hong_Kong.html}
}
\usage{
hkio_users_df

hkio_tweets_df1

hkio_tweets_df2

hkio_tweets_df3

hkio_tweets_df4

hkio_tweets_df5

hkio_tweets_df6

hkio_tweets_df7

hkio_tweets_df8

hkio_tweets_df9
}
\description{
2019 Twitter Information Operations in Hong Kong
}
\details{
On 19 August 2019, Twitter disclosed a significant state-sponsored
information operation targeting Hong Kong, specifically the protest movement.
}
\seealso{
\code{\link[=hkio_all_tweets]{hkio_all_tweets()}}
}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
\keyword{datasets}
