% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hkio-data.R, R/hkio_all_tweets.R
\docType{data}
\name{hkio_users_df}
\alias{hkio_users_df}
\alias{hkio_all_tweets}
\title{hkio}
\format{Ten \code{data.table}s. One for users, and nine covering the more than 3.6 million tweets.}
\source{
\url{https://blog.twitter.com/en_us/topics/company/2019/information_operations_directed_at_Hong_Kong.html}
}
\usage{
hkio_users_df

hkio_all_tweets()
}
\value{
\code{data.table}
}
\description{
2019 Twitter Information Operations in Hong Kong
}
\details{
On 19 August 2019, Twitter disclosed a significant state-sponsored
information operation targeting Hong Kong, specifically the protest movement.
}
\examples{
all_tweets <- hkio_all_tweets()

}
\author{
Brendan Knapp \email{brendan.g.knapp@gmail.com}

Brendan Knapp \email{brendan.g.knapp@gmail.com}
}
\keyword{datasets}
